add back `madara_load_more`
