+++ src/en/setsuscans/src/eu/kanade/tachiyomi/extension/en/setsuscans/SetsuScans.kt
@@ -2,7 +2,6 @@ package eu.kanade.tachiyomi.extension.en.setsuscans
 
 import eu.kanade.tachiyomi.multisrc.madara.Madara
 import eu.kanade.tachiyomi.network.interceptor.rateLimit
-import eu.kanade.tachiyomi.source.model.MangasPage
 import okhttp3.Interceptor
 import okhttp3.Response
 
@@ -60,23 +59,7 @@ class SetsuScans : Madara(
     }
 
     override val useNewChapterEndpoint = true
-
-    override fun popularMangaParse(response: Response) =
-        super.popularMangaParse(response).fixNextPage()
-
-    override fun latestUpdatesParse(response: Response) =
-        super.latestUpdatesParse(response).fixNextPage()
-
-    override fun searchMangaParse(response: Response) =
-        super.searchMangaParse(response).fixNextPage()
-
-    private fun MangasPage.fixNextPage(): MangasPage {
-        return if (mangas.size < 12) {
-            MangasPage(mangas, false)
-        } else {
-            this
-        }
-    }
+    override val useLoadMoreRequest = true
 
     override val mangaDetailsSelectorStatus = "div.summary-heading:contains(status) + div.summary-content"
 }
